fix issues in Frontend source code

+++ src/components/modals/CreateOrUpdateTechModal.vue
@@ -6,6 +6,7 @@
     import Input from '@components/ui/input/Input.vue';
 
     import type { Tech as TechType } from '@type/entities';
+    import type { FormErrors } from "@type/responses"
 
     import { createTech, updateTech, deleteTech } from '@services/techService';
 
@@ -23,7 +24,10 @@
 
     const techs = defineModel<TechType[]>()
 
-    const errors = ref<any>()
+    const isSubmitLoading = ref(false)
+    const isDeleteLoading = ref(false)
+
+    const errors = ref<FormErrors>({})
 
     const currentTech = ref<TechType>({
         id: '',
@@ -37,19 +41,21 @@
     })
 
     const handleSubmit = async ()=>{
-        errors.value = null
+        if (isSubmitLoading.value) return
+        
 
-        if(props.isEditing){
-            
-            const updatePayload:{name: string, icon?: File} =  {
-                name: techForm.value.name
-            }
-            if(techForm.value.icon){
-                updatePayload.icon = techForm.value.icon
-            }
-            try {
+        try {
+            isSubmitLoading.value = true
+            errors.value = {}
+            if(props.isEditing){
+                const updatePayload:{name: string, icon?: File} =  {
+                    name: techForm.value.name
+                }
+                if(techForm.value.icon){
+                    updatePayload.icon = techForm.value.icon
+                }
                 const res = await updateTech(currentTech.value.id, updatePayload)
-                
+            
                 const updatedTech = res.data
 
                 const index = techs.value?.findIndex(
@@ -62,17 +68,9 @@
                 ) {
                     techs.value[index] = updatedTech
                 }
-
                 toast.success(res.message)
-                emit('close')
-            } catch (error: any) {
-                errors.value = error.response?.data?.errors
-                toast.error(error.response?.data?.message)
-            }
-        }else{
-            try {
+            }else{
                 if (!techForm.value.icon){
-
                     toast.error("Icon is required")
                     return
                 }
@@ -82,18 +80,22 @@
                 })
                 
                 techs.value?.push(res.data)
-
                 toast.success(res.message)
-                emit('close')
-            } catch (error: any) {
-                errors.value = error.response?.data?.errors
-                toast.error(error.response?.data?.message)
             }
+            emit('close')
+        } catch (error: any) {
+            errors.value = error.response?.data?.errors
+            toast.error(error.response?.data?.message)
+        }finally {
+            isSubmitLoading.value = false
         }
     }
 
     const handleDelete = async ()=>{
+        if (isDeleteLoading.value) return
+
         try {
+            isDeleteLoading.value = true
             const res = await deleteTech(currentTech.value.id)
 
             techs.value = techs.value?.filter(
@@ -104,6 +106,8 @@
             emit('close')
         } catch (error:any) {
             toast.error(error.response?.data?.message)
+        }finally {
+            isDeleteLoading.value = false
         }
     }
 
@@ -113,7 +117,7 @@
             if (isEditing &&  tech) {
                 currentTech.value = { ...tech }
                 techForm.value.name = tech.name
-                errors.value = ""
+                errors.value = {}
             }else {
                 currentTech.value = {
                     id: '',
@@ -124,7 +128,7 @@
                     name: '',
                     icon: undefined
                 }
-                errors.value = ""
+                errors.value = {}
             }
         },
         { immediate: true }
@@ -134,12 +138,16 @@
 <template>
     <BaseModal :open="open" @close="emit('close')">
         <h1 style="text-align: center;">{{ isEditing ? "Update Tech" : "Create new tech" }}</h1>
-        <Input :error="errors?.name" label="Name" placeholder="Java" v-model="techForm.name"/>
-        <Input :error="errors?.icon" label="Icon Image" v-model="techForm.icon" type="file"/>
+        <Input :error="errors?.name?.[0]" label="Name" placeholder="Java" v-model="techForm.name"/>
+        <Input :error="errors?.icon?.[0]" label="Icon Image" v-model="techForm.icon" type="file"/>
         <img style="width: 60px; object-fit: contain;" :src="currentTech.icon" alt="">
         <div style="display: flex; gap:15px; align-items: center;">
-            <button class="btn" @click="handleSubmit">Submit</button>
-            <button v-if="isEditing" class="btn" @click="handleDelete">Delete</button>
+            <button class="btn" :disabled="isSubmitLoading" @click="handleSubmit">
+                {{ isSubmitLoading ? "Loading..." : "Submit" }}
+            </button>
+            <button v-if="isEditing" class="btn" :disabled="isDeleteLoading" @click="handleDelete">
+                {{ isDeleteLoading ? "Deleting..." : "Delete" }}
+            </button>
         </div>
     </BaseModal>
 </template>
