+++ src/components/modals/UpdateOrCreateWorkExperienceModal.vue
@@ -6,6 +6,7 @@
     import Input from '@components/ui/input/Input.vue';
 
     import type { UserWorkExperience } from '@type/entities';
+    import type { FormErrors } from "@type/responses"
 
     import { createWorkExperience, updateWorkExperience, deleteWorkExperience } from '@services/workExperienceService';
 
@@ -21,9 +22,12 @@
 
     const toast = useToast()
 
+    const isSubmitLoading = ref(false)
+    const isDeleteLoading = ref(false)
+
     const workExperiences = defineModel<UserWorkExperience[]>()
 
-    const errors = ref<any>()
+    const errors = ref<FormErrors>()
 
     const currentWorkExperience = ref<UserWorkExperience>({
         id: '',
@@ -35,10 +39,13 @@
     })
 
     const handleSubmit = async ()=>{
-        errors.value = null
+        if (isSubmitLoading.value) return
 
-        if(props.isEditing){
-            try {
+        try {
+            isSubmitLoading.value = true
+            errors.value = {}
+
+            if(props.isEditing){
                 const { id, userId, ...workExperiencePayload } = currentWorkExperience.value 
                 const res = await updateWorkExperience(currentWorkExperience.value.id, workExperiencePayload)
                 
@@ -56,29 +63,29 @@
                 }
 
                 toast.success(res.message)
-                emit('close')
-            } catch (error: any) {
-                errors.value = error.response?.data?.errors
-                toast.error(error.response?.data?.message)
-            }
-        }else{
-            try {
+            }else{
                 const { id, userId, ...workExperiencePayload } = currentWorkExperience.value 
                 const res = await createWorkExperience(workExperiencePayload)
                 
                 workExperiences.value?.push(res.data)
 
                 toast.success(res.message)
-                emit('close')
-            } catch (error: any) {
-                errors.value = error.response?.data?.errors
-                toast.error(error.response?.data?.message)
             }
+            emit('close')
+        } catch (error: any) {
+            errors.value = error.response?.data?.errors
+            toast.error(error.response?.data?.message)
+        }finally{
+            isSubmitLoading.value = false
         }
     }
 
     const handleDelete = async ()=>{
+        if (isDeleteLoading.value) return
+
         try {
+            isDeleteLoading.value = true
+
             const res = await deleteWorkExperience(currentWorkExperience.value.id)
 
             workExperiences.value = workExperiences.value?.filter(
@@ -89,6 +96,8 @@
             emit('close')
         } catch (error:any) {
             toast.error(error.response?.data?.message)
+        }finally{
+            isDeleteLoading.value = false
         }
     }
 
@@ -97,7 +106,7 @@
         ([isEditing, workExperience]) => {
             if (isEditing &&  workExperience) {
                 currentWorkExperience.value = { ...workExperience }
-                errors.value = ""
+                errors.value = {}
             }else {
                 currentWorkExperience.value = {
                     id: '',
@@ -107,7 +116,7 @@
                     endDate: '',
                     userId: ''
                 }
-                errors.value = ""
+                errors.value = {}
             }
         },
         { immediate: true }
@@ -117,13 +126,17 @@
 <template>
     <BaseModal :open="open" @close="emit('close')">
         <h1 style="text-align: center;">{{ isEditing ? "Update Work Experience" : "Create Work Experience" }}</h1>
-        <Input label="Position" placeholder="Backend developer" v-model="currentWorkExperience.position"/>
-        <Input label="Organization" placeholder="Hodford" v-model="currentWorkExperience.organization"/>
-        <Input label="Start date" placeholder="11/11/2021" v-model="currentWorkExperience.startDate" type="date"/>
-        <Input label="End date" placeholder="11/12/2021" v-model="currentWorkExperience.endDate" type="date"/>
+        <Input :error="errors?.position?.[0]" label="Position" placeholder="Backend developer" v-model="currentWorkExperience.position"/>
+        <Input :error="errors?.organization?.[0]" label="Organization" placeholder="Hodford" v-model="currentWorkExperience.organization"/>
+        <Input :error="errors?.startDate?.[0]" label="Start date" placeholder="11/11/2021" v-model="currentWorkExperience.startDate" type="date"/>
+        <Input :error="errors?.endDate?.[0]" label="End date" placeholder="11/12/2021" v-model="currentWorkExperience.endDate" type="date"/>
         <div style="display: flex; gap:15px; align-items: center;">
-            <button class="btn" @click="handleSubmit">Submit</button>
-            <button v-if="isEditing" class="btn" @click="handleDelete">Delete</button>
+            <button class="btn" :disabled="isSubmitLoading" @click="handleSubmit">
+                {{ isSubmitLoading ? "Loading..." : "Submit" }}
+            </button>
+            <button v-if="isEditing" :disabled="isDeleteLoading" class="btn" @click="handleDelete">
+                {{ isDeleteLoading ? "Deleting..." : "Delete" }}
+            </button>
         </div>
     </BaseModal>
 </template>
